Sensor Server/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.0.0
